+++ http4k-contract/src/main/kotlin/org/http4k/contract/openapi/v3/model.kt
@@ -25,7 +25,7 @@ data class ApiPath<NODE>(
     val description: String?,
     val tags: List<String>,
     val parameters: List<RequestParameter<NODE>>,
-    val requestBody: RequestContents<NODE>,
+    val requestBody: RequestContents<NODE>?,
     val responses: Map<String, ResponseContents<NODE>>,
     val security: NODE,
     val operationId: String
@@ -33,7 +33,7 @@ data class ApiPath<NODE>(
     fun definitions() = listOfNotNull(
         responses.flatMap { it.value.definitions() },
         parameters.filterIsInstance<HasSchema<NODE>>().flatMap { it.definitions() },
-        requestBody.definitions().toList()
+        requestBody?.definitions()?.toList()
     ).flatten()
 }
 
@@ -60,12 +60,13 @@ sealed class BodyContent {
     }
 }
 
-class RequestContents<NODE>(val content: Map<String, BodyContent> = emptyMap()) : HasSchema<NODE> {
-    override fun definitions() = content.values
-        .filterIsInstance<HasSchema<NODE>>()
-        .flatMap { it.definitions() }
+class RequestContents<NODE>(val content: Map<String, BodyContent>? = null) : HasSchema<NODE> {
+    override fun definitions() = content?.values
+        ?.filterIsInstance<HasSchema<NODE>>()
+        ?.flatMap { it.definitions() }
+        ?: emptyList()
 
-    val required = content.isNotEmpty()
+    val required = content != null
 }
 
 class ResponseContents<NODE>(val description: String?, val content: Map<String, BodyContent> = emptyMap()) : HasSchema<NODE> {